# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AssignCategoriesRequest:
    properties:
      category_ids:
        items:
          type: string
        minItems: 0
        type: array
    required:
    - category_ids
    type: object
  dto.CategoryResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/types.Metadata'
      name:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/types.Status'
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  dto.CreatePlaceImageRequest:
    properties:
      alt:
        maxLength: 255
        type: string
      metadata:
        $ref: '#/definitions/types.Metadata'
      pos:
        maximum: 100
        minimum: 0
        type: integer
      url:
        maxLength: 500
        type: string
    required:
    - url
    type: object
  dto.CreatePlaceRequest:
    properties:
      address:
        additionalProperties:
          type: string
        type: object
      location:
        $ref: '#/definitions/types.GeoPoint'
      long_description:
        maxLength: 10000
        type: string
      place_type:
        $ref: '#/definitions/types.PlaceType'
      primary_image_url:
        maxLength: 500
        type: string
      short_description:
        maxLength: 1000
        type: string
      slug:
        maxLength: 100
        minLength: 3
        type: string
      subtitle:
        maxLength: 500
        type: string
      thumbnail_url:
        maxLength: 500
        type: string
      title:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - location
    - place_type
    - slug
    - title
    type: object
...
  dto.FeedRequest:
    properties:
      end_time:
        type: string
      expand:
        type: string
      limit:
        maximum: 1000
        minimum: 1
        type: integer
      min_rating_avg:
        type: number
      min_rating_count:
        description: Optional quality gate (global; section can override)
        minimum: 0
        type: integer
      offset:
        minimum: 0
        type: integer
      order:
        enum:
        - asc
        - desc
        type: string
      sections:
        items:
          $ref: '#/definitions/dto.FeedSectionRequest'
        maxItems: 10
        minItems: 1
        type: array
      sort:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/types.Status'
    required:
    - sections
    type: object
  dto.FeedResponse:
    properties:
      sections:
        items:
          $ref: '#/definitions/dto.FeedSectionResponse'
        type: array
    type: object
  dto.FeedSectionRequest:
    properties:
      center:
        allOf:
        - $ref: '#/definitions/types.GeoPoint'
        description: Geospatial fields (for nearby section)
      end_time:
        type: string
      expand:
        type: string
      limit:
        maximum: 1000
        minimum: 1
        type: integer
      min_rating_avg:
        type: number
      min_rating_count:
        description: Optional quality gate (section-level; overrides global)
        minimum: 0
        type: integer
      offset:
        minimum: 0
        type: integer
      order:
        enum:
        - asc
        - desc
        type: string
      radius_km:
        default: 5
        maximum: 50
        minimum: 0.1
        type: number
      sort:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/types.Status'
      type:
        $ref: '#/definitions/types.FeedSectionType'
    required:
    - type
    type: object
  dto.FeedSectionResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.PlaceResponse'
        type: array
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
      type:
        $ref: '#/definitions/types.FeedSectionType'
    type: object
...
  dto.ListPlacesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.PlaceResponse'
        type: array
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
    type: object
...
  dto.PlaceImageResponse:
    properties:
      alt:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/types.Metadata'
      place_id:
        type: string
      pos:
        type: integer
      status:
        $ref: '#/definitions/types.Status'
      updated_at:
        type: string
      updated_by:
        type: string
      url:
        type: string
    type: object
  dto.PlaceResponse:
    properties:
      address:
        additionalProperties:
          type: string
        type: object
      categories:
        items:
          $ref: '#/definitions/dto.CategoryResponse'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/dto.PlaceImageResponse'
        type: array
      location:
        $ref: '#/definitions/types.GeoPoint'
      long_description:
        type: string
      place_type:
        $ref: '#/definitions/types.PlaceType'
      popularity_score:
        type: number
      primary_image_url:
        type: string
      rating_avg:
        type: number
      rating_count:
        type: integer
      short_description:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/types.Status'
      subtitle:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      view_count:
        description: Engagement fields for feed functionality
        type: integer
    type: object
...
  dto.UpdatePlaceImageRequest:
    properties:
      alt:
        type: string
      metadata:
        $ref: '#/definitions/types.Metadata'
      pos:
        minimum: 0
        type: integer
      url:
        type: string
    type: object
  dto.UpdatePlaceRequest:
    properties:
      address:
        additionalProperties:
          type: string
        type: object
      location:
        $ref: '#/definitions/types.GeoPoint'
      long_description:
        maxLength: 10000
        type: string
      primary_image_url:
        maxLength: 500
        type: string
      short_description:
        maxLength: 1000
        type: string
      slug:
        maxLength: 100
        minLength: 3
        type: string
      subtitle:
        maxLength: 500
        type: string
      thumbnail_url:
        maxLength: 500
        type: string
      title:
        maxLength: 255
        minLength: 2
        type: string
    type: object
...
  ierr.ErrorDetail:
    properties:
      details:
        additionalProperties: {}
        type: object
      internal_error:
        type: string
      message:
        type: string
    type: object
  ierr.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/ierr.ErrorDetail'
      success:
        type: boolean
    type: object
...
  types.FeedSectionType:
    enum:
    - latest
    - trending
    - popular
    - nearby
    - discover
    type: string
    x-enum-varnames:
    - SectionTypeLatest
    - SectionTypeTrending
    - SectionTypePopular
    - SectionTypeNearby
    - SectionTypeDiscover
  types.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  types.Metadata:
    additionalProperties:
      type: string
    type: object
  types.PaginationResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  types.PlaceType:
    enum:
    - temple
    - restaurant
    - museum
    - park
    - experience
    type: string
    x-enum-varnames:
    - PlaceTypeTemple
    - PlaceTypeRestaurant
    - PlaceTypeMuseum
    - PlaceTypePark
    - PlaceTypeExperience
...
  types.Status:
    enum:
    - published
    - draft
    - archived
    - deleted
    type: string
    x-enum-varnames:
    - StatusPublished
    - StatusDraft
    - StatusArchived
    - StatusDeleted
...
host: 5p9ubi66hh.execute-api.ap-south-1.amazonaws.com
info:
  contact:
    email: support@nashikdarshan.com
    name: API Support
  description: API for Nashik Darshan
  termsOfService: http://nashikdarshan.com/terms/
  title: Nashik Darshan API
  version: "1.0"
paths:
...
  /feed:
    post:
      consumes:
      - application/json
      description: Get feed data with multiple sections (trending, popular, latest,
        nearby)
      parameters:
      - description: dto.FeedRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FeedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FeedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: Get feed data
      tags:
      - Place
...
  /places:
    get:
      consumes:
      - application/json
      description: Get a paginated list of places with filtering and pagination
      parameters:
      - in: query
        name: end_time
        type: string
      - in: query
        name: expand
        type: string
      - in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - in: query
        name: min_rating_avg
        type: number
      - description: Optional quality gate (e.g. min reviews / min rating to appear
          in feed)
        in: query
        minimum: 0
        name: min_rating_count
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: place_types
        type: array
      - description: 'radius in meters (cap: 15km for v1)'
        in: query
        name: radius_m
        type: number
      - description: Search
        in: query
        name: search_query
        type: string
      - collectionFormat: csv
        description: Custom filters
        in: query
        items:
          type: string
        name: slug
        type: array
      - in: query
        name: sort
        type: string
      - in: query
        name: start_time
        type: string
      - enum:
        - published
        - draft
        - archived
        - deleted
        in: query
        name: status
        type: string
        x-enum-varnames:
        - StatusPublished
        - StatusDraft
        - StatusArchived
        - StatusDeleted
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListPlacesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: List places
      tags:
      - Place
    post:
      consumes:
      - application/json
      description: Create a new place with the provided details
      parameters:
      - description: dto.CreatePlaceRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePlaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PlaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Create a new place
      tags:
      - Place
  /places/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a place
      parameters:
      - description: Place ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Delete a place
      tags:
      - Place
    get:
      consumes:
      - application/json
      description: Get a place by its ID
      parameters:
      - description: Place ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlaceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: Get place by ID
      tags:
      - Place
    put:
      consumes:
      - application/json
      description: Update an existing place
      parameters:
      - description: Place ID
        in: path
        name: id
        required: true
        type: string
      - description: dto.UpdatePlaceRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePlaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Update a place
      tags:
      - Place
  /places/{id}/categories:
    put:
      consumes:
      - application/json
      description: Assign categories to a place by replacing existing category relationships
      parameters:
      - description: Place ID
        in: path
        name: id
        required: true
        type: string
      - description: dto.AssignCategoriesRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AssignCategoriesRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Assign categories to a place
      tags:
      - Place
  /places/{id}/images:
    get:
      consumes:
      - application/json
      description: Get all images for a place
      parameters:
      - description: Place ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PlaceImageResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: Get place images
      tags:
      - Place
    post:
      consumes:
      - application/json
      description: Add an image to a place
      parameters:
      - description: Place ID
        in: path
        name: id
        required: true
        type: string
      - description: dto.CreatePlaceImageRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePlaceImageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PlaceImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Add image to place
      tags:
      - Place
  /places/{id}/view:
    post:
      consumes:
      - application/json
      description: Increment the view count for a specific place
      parameters:
      - description: Place ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: Increment view count for a place
      tags:
      - Place
  /places/images/{image_id}:
    delete:
      consumes:
      - application/json
      description: Delete a place image
      parameters:
      - description: Image ID
        in: path
        name: image_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Delete place image
      tags:
      - Place
    put:
      consumes:
      - application/json
      description: Update an existing place image
      parameters:
      - description: Image ID
        in: path
        name: image_id
        required: true
        type: string
      - description: dto.UpdatePlaceImageRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePlaceImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlaceImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Update place image
      tags:
      - Place
  /places/refresh-scores:
    post:
      consumes:
      - application/json
      description: Recalculates popularity_score for all places; for use by cron.
        Requires authentication.
      produces:
      - application/json
      responses:
        "200":
          description: 'e.g. {\"places_updated\": N}'
          schema:
            additionalProperties:
              type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Refresh place scores
      tags:
      - Place
  /places/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get a place by its slug
      parameters:
      - description: Place slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlaceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: Get place by slug
      tags:
      - Place
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key for server-to-server or script access. Use X-API-Key header
      (alternative to Bearer).
    in: header
    name: X-API-Key
    type: apiKey
  Authorization:
    description: JWT Bearer token, e.g. `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"